Okay so here's what happened: http://stackoverflow.com/questions/25012587/creating-variables-within-android-services (I posted this Q on stack overflow).
The people on SO can be kinda mean.

Other strange bugs: When using Notification Builder, you need to set a SmallIcon first which makes no sense.

But yeah. Notification thing works now, Snippet class is set up correctly, and things look good.

Next step is to save to an ArrayList or other data structure after figuring out if we can use a intermediary or singleton class for data. Then figure out dumping to disk, and then follow the Google Copy Paste API guide to parse snippets appropriately within the app's history thing. Finally implement the pasting mechanism. For that we need to decide if we want another Service or if we want to build on this one (everything points to another one except the fact that we *may* need a separate notification which is a deal breaker).

rbizward :D