I hate UI work.
I got a primitive version of the trigger for the paste thing working. Drag down from the top left corner to see something...

Need to change it so that it shows the actual paste UI (to be created) on drag and then on release selects the text below (need to find a good way to do this - not sure if Views are capable, or if they are even given touch data)

Also need to not hard code values and switch over to dp...

Also primes.