Couple of Development Process Changes: First off, Crashlytics has been added to the project. Next Build (with Gradle) has been fixed. Next, options have been added to the Release BuildType to configure uploading directly to Crashlytics after a build. Next, a keystore and a key have been added to the project and the appropriate Gradle config files changes have been made so that both debug and release builds are signed and signed with the same key. (Crashlytics needed this, and it also allows app upgrading to actually happen properly. Next up is a tiny change to the .travis.yml file: the github OAuth key is now sourced from a variable set securely by Travis CI so that we don't have to commit important API keys.
Finally, to make continuous integration tools able to sign things as well as Android Studio the following has been implemented:
The Gradle signing configuration has been set to use environment specific variables for the password, etc. These passwords are set into the environment by a script that WILL NOT be committed to this repo (others working on this project can ask me for this script). Gradle will call this script if it exists. Travis has these environment variables set in its settings. TODO: Look into encrypted variables for Travis.
Additionally the keystore has been committed to the Github Repo so that it is available to the local builds and to continuous integration tools.

UPDATE:
This commit was held off because the set up for using variables for the password of the keystore and the key file was incomplete. Rather than risk committin potentially important passwords to a public Github repo, the commit was held back:
Here are the changes that have been made since the last commit:
The continuous integration builds and the local builds have been given different buildTypes and different correpsonding signing setups and here's why:
The local build uses variables set in the user's gradle.properties file (in their home directory) rather than variables set in a script as this is better for cross platform support and those lost helpless souls still using Windows. Again contact me for the passwords.

That's all. Nothing has been tested, which means that probably nothing works...